pyproject.toml
[project]
name = "data-engineer-case"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flytekit>=1.15.3",
    "flytekitplugins-spark>=1.15.3",
    "psycopg2-binary>=2.9.10",
    "pyspark>=3.5.5",
    "setuptools>=78.1.0",
]

[dependency-groups]
dev = ["pre-commit>=4.2.0", "pytest>=8.3.5"]
